Refactor
Change tests to separate files
Updated templates
Documentation updates
