+++ .gitattributes
@@ -7,12 +7,12 @@
 /.gitattributes             export-ignore
 /.gitignore                 export-ignore
 /.php-cs-fixer.dist.php     export-ignore
+/Dockerfile                 export-ignore
+/docker-compose.yml         export-ignore
+/Makefile                   export-ignore
 /phpunit.xml.dist           export-ignore
 /RELEASE_CHECKLIST.md       export-ignore
 # directory ignores
 /.github                    export-ignore
-/tests                      export-ignore
-/Dockerfile                 export-ignore
-/docker-compose.yml         export-ignore
-/Makefile                   export-ignore
 /build                      export-ignore
+/tests                      export-ignore
