+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,8 +1,8 @@
 ---
-name: Bug report
-about: Create a report to help us improve
+name: "\U0001F41B Bug Report"
+about: "If something isn't working as expected \U0001F914."
 title: ''
-labels: ''
+labels: bug, triage
 assignees: ''
 
 ---
@@ -13,10 +13,13 @@ A clear and concise description of what the bug is.
 **To Reproduce**
 Steps to reproduce the behavior:
 
-1. Go to '...'
-2. Click on '....'
-3. Scroll down to '....'
-4. See error
+What version are you using?
+
+Example code to reproduce
+<!-- This is a public repo, DO NOT ADD SECRETS TO EXAMPLE CODE -->
+```php
+$your = 'code' . 'here';
+```
 
 **Expected behavior**
 A clear and concise description of what you expected to happen.
@@ -24,11 +27,9 @@ A clear and concise description of what you expected to happen.
 **Screenshots**
 If applicable, add screenshots to help explain your problem.
 
-**Desktop (please complete the following information):**
-
-- OS: [e.g. iOS]
-- Browser [e.g. chrome, safari]
-- Version [e.g. 22]
-
 **Additional context**
 Add any other context about the problem here.
+
+**Environment**
+- php version(s): [e.g. 5.6.7, v8.4.0-beta.34]
+- OS: [e.g. OSX 10.13.4, Ubuntu 18.04]
