+++ .gitignore
@@ -6,3 +6,4 @@ composer.phar
 composer.lock
 phpunit.xml
 .php_cs.cache
+docker-compose.override.yml
